Revert "8c: remaining long -> int32"
This reverts commit ee9d700b256fbf4318dc5b083446a957f97f3814.

+++ compilers/8c/gc.h
@@ -53,7 +53,7 @@ struct	Prog
 	Adr	from;
 	Adr	to;
 	Prog*	link;
-	int32	lineno;
+	long	lineno;
 	short	as;
 };
 #define	P	((Prog*)0)
@@ -62,7 +62,7 @@ struct	Case
 {
 	Case*	link;
 	vlong	val;
-	int32	label;
+	long	label;
 	char	def;
 	char isv;
 };
@@ -71,12 +71,12 @@ struct	Case
 struct	C1
 {
 	vlong	val;
-	int32	label;
+	long	label;
 };
 
 struct	Var
 {
-	int32	offset;
+	long	offset;
 	Sym*	sym;
 	char	name;
 	char	etype;
@@ -84,8 +84,8 @@ struct	Var
 
 struct	Reg
 {
-	int32	pc;
-	int32	rpo;		/* reverse post ordering */
+	long	pc;
+	long	rpo;		/* reverse post ordering */
 
 	Bits	set;
 	Bits	use1;
@@ -98,11 +98,11 @@ struct	Reg
 	Bits	regdiff;
 	Bits	act;
 
-	int32	regu;
-	int32	loop;		/* could be shorter */
+	long	regu;
+	long	loop;		/* could be shorter */
 
 	Reg*	log5;
-	int32	active;
+	long	active;
 
 	Reg*	p1;
 	Reg*	p2;
@@ -131,25 +131,25 @@ struct	Rgn
 	short	regno;
 };
 
-EXTERN	int32	breakpc;
-EXTERN	int32	nbreak;
+EXTERN	long	breakpc;
+EXTERN	long	nbreak;
 EXTERN	Case*	cases;
 EXTERN	Node	constnode;
 EXTERN	Node	fconstnode;
-EXTERN	int32	continpc;
-EXTERN	int32	curarg;
-EXTERN	int32	cursafe;
+EXTERN	long	continpc;
+EXTERN	long	curarg;
+EXTERN	long	cursafe;
 EXTERN	Prog*	firstp;
 EXTERN	Prog*	lastp;
-EXTERN	int32	maxargsafe;
+EXTERN	long	maxargsafe;
 EXTERN	int	mnstring;
 EXTERN	Node*	nodrat;
 EXTERN	Node*	nodret;
 EXTERN	Node*	nodsafe;
-EXTERN	int32	nrathole;
-EXTERN	int32	nstring;
+EXTERN	long	nrathole;
+EXTERN	long	nstring;
 EXTERN	Prog*	p;
-EXTERN	int32	pc;
+EXTERN	long	pc;
 EXTERN	Node	regnode;
 EXTERN	Node	fregnode0;
 EXTERN	Node	fregnode1;
@@ -158,8 +158,8 @@ EXTERN	Sym*	symrathole;
 EXTERN	Node	znode;
 EXTERN	Prog	zprog;
 EXTERN	int	reg[D_NONE];
-EXTERN	int32	exregoffset;
-EXTERN	int32	exfregoffset;
+EXTERN	long	exregoffset;
+EXTERN	long	exfregoffset;
 
 #define	BLOAD(r)	band(bnot(r->refbehind), r->refahead)
 #define	BSTORE(r)	band(bnot(r->calbehind), r->calahead)
@@ -183,8 +183,8 @@ EXTERN	Bits	params;
 EXTERN	Bits	consts;
 EXTERN	Bits	addrs;
 
-EXTERN	int32	regbits;
-EXTERN	int32	exregbits;
+EXTERN	long	regbits;
+EXTERN	long	exregbits;
 
 EXTERN	int	change;
 EXTERN	int	suppress;
@@ -194,9 +194,9 @@ EXTERN	Reg*	lastr;
 EXTERN	Reg	zreg;
 EXTERN	Reg*	freer;
 EXTERN	Var	var[NVAR];
-EXTERN	int32*	idom;
+EXTERN	long*	idom;
 EXTERN	Reg**	rpo2r;
-EXTERN	int32	maxnr;
+EXTERN	long	maxnr;
 
 extern	char*	anames[];
 
@@ -220,7 +220,7 @@ void	reglcgen(Node*, Node*, Node*);
 void	lcgen(Node*, Node*);
 void	bcgen(Node*, int);
 void	boolgen(Node*, int, Node*);
-void	sugen(Node*, Node*, int32);
+void	sugen(Node*, Node*, long);
 int	needreg(Node*, int);
 
 /*
@@ -239,7 +239,7 @@ void	gclean(void);
 void	nextpc(void);
 void	gargs(Node*, Node*, Node*);
 void	garg1(Node*, Node*, Node*, int, Node**);
-Node*	nodconst(int32);
+Node*	nodconst(long);
 int	nareg(int);
 Node*	nodfconst(double);
 int	nodreg(Node*, Node*, int);
@@ -260,7 +260,7 @@ void	fgopcode(int, Node*, Node*, int, int);
 void	gopcode(int, Type*, Node*, Node*);
 int	samaddr(Node*, Node*);
 void	gbranch(int);
-void	patch(Prog*, int32);
+void	patch(Prog*, long);
 int	sconst(Node*);
 void	gpseudo(int, Sym*, Node*);
 
@@ -269,13 +269,13 @@ void	gpseudo(int, Sym*, Node*);
  */
 int	swcmp(const void*, const void*);
 void	doswit(Node*);
-void	swit1(C1*, int, int32, Node*);
+void	swit1(C1*, int, long, Node*);
 void	casf(void);
 void	bitload(Node*, Node*, Node*, Node*, Node*);
 void	bitstore(Node*, Node*, Node*, Node*, Node*);
-int32	outstring(char*, int32);
+long	outstring(char*, long);
 void	nullwarn(Node*, Node*);
-void	gextern(Sym*, Node*, int32, int32);
+void	gextern(Sym*, Node*, long, long);
 void	outcode(void);
 void	ieeedtod(Ieee*, double);
 
@@ -300,12 +300,12 @@ void	regopt(Prog*);
 void	addmove(Reg*, int, int, int);
 Bits	mkvar(Reg*, Adr*, int);
 void	prop(Reg*, Bits, Bits);
-void	loopit(Reg*, int32);
+void	loopit(Reg*, long);
 void	synch(Reg*, Bits);
-uint32	allreg(uint32, Rgn*);
+ulong	allreg(ulong, Rgn*);
 void	paint1(Reg*, int);
-uint32	paint2(Reg*, int);
-void	paint3(Reg*, int, int32, int);
+ulong	paint2(Reg*, int);
+void	paint3(Reg*, int, long, int);
 void	addreg(Adr*, int);
 
 /*
@@ -327,10 +327,10 @@ int	copyau(Adr*, Adr*);
 int	copysub(Adr*, Adr*, Adr*, int);
 int	copysub1(Prog*, Adr*, Adr*, int);
 
-int32	RtoB(int);
-int32	FtoB(int);
-int	BtoR(int32);
-int	BtoF(int32);
+long	RtoB(int);
+long	FtoB(int);
+int	BtoR(long);
+int	BtoF(long);
 
 #define	D_HI	D_NONE
 #define	D_LO	D_NONE
@@ -342,8 +342,8 @@ int	cond(int);
 int	com64(Node*);
 void	com64init(void);
 void	bool64(Node*);
-int32	lo64v(Node*);
-int32	hi64v(Node*);
+long	lo64v(Node*);
+long	hi64v(Node*);
 Node*	lo64(Node*);
 Node*	hi64(Node*);
 
@@ -352,8 +352,8 @@ Node*	hi64(Node*);
  */
 void	sdivgen(Node*, Node*, Node*, Node*);
 void	udivgen(Node*, Node*, Node*, Node*);
-void	sdiv2(int32, int, Node*, Node*);
-void	smod2(int32, int, Node*, Node*);
+void	sdiv2(long, int, Node*, Node*);
+void	smod2(long, int, Node*, Node*);
 void	mulgen(Type*, Node*, Node*);
 void	genmuladd(Node*, Node*, int, Node*);
 void	shiftit(Type*, Node*, Node*);
